+++ src/Shardis.Query/Execution/FailureHandling/FailureHandlingExecutor.cs
@@ -1,5 +1,3 @@
-using System.Threading;
-
 using Shardis.Query.Execution.FailureHandling;
 
 namespace Shardis.Query.Execution;