+++ test/Shardis.Query.Tests/QueryLatencyOpenTelemetryTests.cs
@@ -9,7 +9,6 @@ using Shardis.Model;
 using Shardis.Query.Diagnostics;
 using Shardis.Query.EntityFrameworkCore;
 using Shardis.Query.EntityFrameworkCore.Execution;
-using Shardis.Query.Execution;
 
 namespace Shardis.Query.Tests;